refactor: #714 토론방 FAB 컴포저블을 component 패키지로 이동
